+++ compio-http/src/service.rs
@@ -10,7 +10,9 @@ use send_wrapper::SendWrapper;
 
 use crate::{HttpStream, TlsBackend};
 
-#[derive(Debug, Clone)]
+/// An executor service based on [`compio_runtime`]. It uses
+/// [`compio_runtime::spawn`] interally.
+#[derive(Debug, Default, Clone)]
 pub struct CompioExecutor;
 
 impl Executor<Pin<Box<dyn Future<Output = ()> + Send>>> for CompioExecutor {
@@ -19,12 +21,17 @@ impl Executor<Pin<Box<dyn Future<Output = ()> + Send>>> for CompioExecutor {
     }
 }
 
+/// An HTTP connector service.
+///
+/// It panics when called in a different thread other than the thread creates
+/// it.
 #[derive(Debug, Clone)]
 pub struct Connector {
     tls: TlsBackend,
 }
 
 impl Connector {
+    /// Creates the connector with specific TLS backend.
     pub fn new(tls: TlsBackend) -> Self {
         Self { tls }
     }
