+++ compio-tls/src/lib.rs
@@ -1,6 +1,7 @@
 //! Async TLS streams.
 
 #![warn(missing_docs)]
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 
 mod adapter;
 mod stream;
